+++ src/game/client/components/menus.cpp
@@ -2344,7 +2344,7 @@ int CMenus::MenuImageScan(const char *pName, int IsDir, int DirType, void *pUser
 	// create gray scale version
 	unsigned char *pData = static_cast<unsigned char *>(Info.m_pData);
 	const size_t Step = Info.PixelSize();
-	for(int i = 0; i < Info.m_Width * Info.m_Height; i++)
+	for(size_t i = 0; i < Info.m_Width * Info.m_Height; i++)
 	{
 		int v = (pData[i * Step] + pData[i * Step + 1] + pData[i * Step + 2]) / 3;
 		pData[i * Step] = v;
